services/user-service/requirements.txt
aiokafka==0.12.0 ; python_version >= "3.12" and python_version < "4.0"
aiosmtplib==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
alembic==1.14.0 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.7.0 ; python_version >= "3.12" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.30.0 ; python_version >= "3.12" and python_version < "4.0"
bcrypt==4.2.1 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.12.14 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
dishka==1.4.2 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.12" and python_version < "4.0"
ecdsa==0.19.0 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
fast-depends==2.4.12 ; python_version >= "3.12" and python_version < "4.0"
fastapi-cli[standard]==0.0.7 ; python_version >= "3.12" and python_version < "4.0"
fastapi[all]==0.115.6 ; python_version >= "3.12" and python_version < "4.0"
faststream[kafka]==0.5.33 ; python_version >= "3.12" and python_version < "4.0"
greenlet
#greenlet==3.1.1 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.5 ; python_version >= "3.12" and python_version < "4.0"
mako==1.3.8 ; python_version >= "3.12" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.10.12 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.12" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.5.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.27.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic-extra-types==2.10.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.7.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.10.4 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.12" and python_version < "4.0"
pytest-asyncio==0.25.0 ; python_version >= "3.12" and python_version < "4.0"
pytest==8.3.4 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-jose==3.3.0 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.20 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
rich-toolkit==0.12.0 ; python_version >= "3.12" and python_version < "4.0"
rich==13.9.4 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
shellingham==1.5.4 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.36 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.41.3 ; python_version >= "3.12" and python_version < "4.0"
typer==0.15.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
ujson==5.10.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn[standard]==0.34.0 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.21.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.12" and python_version < "4.0"
watchfiles==1.0.3 ; python_version >= "3.12" and python_version < "4.0"
websockets==14.1 ; python_version >= "3.12" and python_version < "4.0"
